requirements.txt
asgiref==3.7.2
Babel==2.14.0
Django==5.0.2
django-dotenv==1.4.2
django-money==3.4.1
django-phonenumber-field==7.3.0
django-phonenumbers==1.0.1
django-upgrade==1.16.0
djangorestframework==3.14.0
phonenumbers==8.13.30
psycopg2-binary==2.9.9
py-moneyed==3.0
pytz==2024.1
sqlparse==0.4.4
tokenize-rt==5.2.0
typing_extensions==4.9.0
whitenoise==6.6.0
